pyproject.toml
[tool.poetry]
name = "gitflow-pyproject-version-bumper"
version = "0.1.3"
description = "Automatic pyproject.toml based app version bumper"
authors = ["Grigory Bukovsky <booqoffsky@yandex.ru>"]
readme = "README.md"
license = "MIT"
packages = [{include = "gitflow_pyproject_version_bumper"}]
repository = 'https://github.com/booqoffsky/gitflow-pyproject-version-bumper'
keywords = ['gitflow', 'version', 'bumper', 'auto', 'pyproject', 'poetry', 'updater']

[tool.poetry.dependencies]
python = "^3.8"
GitPython = "^3.1.31"
tomlkit = "^0.11.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
